Fix clear bug in BANCStar example code.
Before (from PDF version):
   057: First Name = First Name
   058: Middle = Middle
   059: Last = Last
   060: Name = Full Name
   061: If Last is Null
   062:   Name = First Name + Middle + Last

note the obvious bug in line 61; clearly, we should be checking if "Name" is
null (that is, "Full Name" was null), not "Last", before reconstructing
a standard "Full Name" based on individual components.

PDF will also need to be updated. I recommend correction fluid and a fine-
tipped marker.